Bump the package and module versions to 4.00
Increase the version number of the package to be larger than any of
the previous version numbers of any of the modules, and change all
modules to use the same version as the overall podlators package.
Switch to a simple decimal version number to avoid complexity with
v-strings and portability to old versions of Perl.