Allow any even-length string as quote characters
Allow any even number of characters to be specified as the quote marks
for Pod::Text and Pod::Man (and the corresponding --quotes options of
pod2text and pod2man), rather than being artificially limited to one-
or two-character quotes.  The first half of the string will be used as
the left quote and the second half as the right quote.  This allows
Unicode characters or groff escapes like \(lq and \(rq to be used.
(Partly addresses #103298)