pod2man --release now requires an argument
[pod2man] In previous versions, the -r or --release option could be
specified without an argument and was interpreted as setting that
value to the empty string.  That never made a great deal of sense, and
the original change to Perl was apparently because no one realized one
could pass the empty string as the argument value.  The argument is
now mandatory, but may be the empty string, which will cause some
*roff implementations to use the system default.

See:

    http://osdir.com/ml/lang.perl.perl5.porters/2004-05/msg00125.html

for the original change that made this argument optional.

Also fix some wording issues around the options setting values for the
.TH header.