Switch back to ExtUtils::MakeMaker as the only build system
Revert the switch to Module::Build as the build system.  This creates
a circular dependency with Module::Build, since it wants a newer
version of Pod::Man than in Perl versions prior to 5.10.1.  Instead,
add the new metadata to Makefile.PL and stick with a single build
system that will also work inside Perl core.