Add more metadata and refactor Makefile.PL
Add homepage information to the CPAN metadata and change the canonical
repository location to GitHub.

Also refactor Makefile.PL to build a hash of parameters to
WriteMakefile and then remove the ones that aren't supported by the
local version of ExtUtils::MakeMaker.  This makes Makefile.PL
perltidy-clean and avoids having things wander too far to the right.