Handle .com endings and skip perlpodstyle in core
Handle scripts ending in .com on VMS systems and don't generate the
man page for perlpodstyle when built as part of Perl core.  These are
hopefully the last two changes required to fully merge with Perl core
without core having to maintain a separate build system.  Thanks,
Craig A. Berry.