Tolerate generating man pages from stdin with no name
[Pod::Man] Partially revert change in 4.00 to require the name option
(--name to pod2man) when generating man pages from standard input.
Historically, pod2man silently tolerated this, and there turned out to
be a lot of software that depended on this, making the change too
disruptive.  Instead, silently set the man page title to STDIN in this
case, but warn about it in the documentation.  (#117990)