Change default Pod::Man outupt to UTF-8
Pod::Man now defaults to UTF-8 output as if the utf8 option or the -u
command-line option to pod2man was given.  There is a new encoding
option (-e or --encoding to pod2man) to change the output encoding to
any encoding recognized by Encode.  It can also be set to the special
value "roff", which requests the old behavior of using character
substitutions and *roff escapes to generate ASCII-only output.  The
utf8 and -u options no longer do anything and are ignored.  (#68741)

This is likely to cause problems on at least some systems, but the
number of problems caused by not supporting Unicode by default seems
higher to me.  Hopefully this will be a forcing function for every
system to handle UTF-8-encoded man pages properly.