Overhaul Pod::Text encoding management
Pod::Text now supports an encoding option, and pod2text supports a -e
or --encoding option, to force the output encoding, similar to
Pod::Man.  The utf8 option is still supported and is equivalent to
setting encoding to UTF-8.

Pod::Text now defaults to UTF-8 encoding if it sees a non-ASCII
character on a non-EBCDIC system and the input encoding is not
specified.  This should fix rendering of E<> escapes of non-ASCII
characters in POD files that don't specify an input encoding, at the
cost of assuming UTF-8 output.  Pod::Text also now commits to an
encoding the first time it outputs a non-ASCII character and sticks
with that encoding for the rest of the file, even if the detected or
declared input encoding changes.