Add groff Pod::Man encoding and make it default
Both groff and mandoc support \[uNNNN] escapes natively, and this
appears to be preferrable to using UTF-8 (which for groff requires
a pass through preconv that just converts it to \[uNNNN] anyway).
Implement this encoding as the groff encoding, and make this the
default instead of UTF-8, since it should be roughly equivalently
portable.