Add encoding tests, switch Pod::Man back to UTF-8
Add a test POD file and various output files to test various
aspects of Unicode encoding for Pod::Man.  Test this on a bunch
of different UNIX platforms and record the results.  Based on that
testing, switch the default back to UTF-8, since it's simpler and
seems to be as portable as groff escapes.