Update Pod::Man documentation
Add a discussion of sentence spacing to the CAVEATS section, since
I don't see a good and maintainable way for Pod::Man to help with
this directly.

Restructure the API documentation for Pod::Man to use the CLASS
METHODS and INSTANCE METHODS sections and document the most
important Pod::Simple methods.

Make some other minor cleanups to the rest of the documentation.