Improve quote escaping in .TH macro
Use a utility function to do proper escaping for macro arguments,
and use it uniformly for the arguments to the .TH macro instead of
rolling multiple, slightly-different versions of the same logic.
This also avoids over-quoting the arguments when this is not needed,
which is too minor to note in Changes but seems cleaner.

Add an additional test for this behavior.