Improve quote handling in Pod::Man
Change LQUOTE and RQUOTE in Pod::Man to be the literal quotes
instead of the escaped quotes, and do the escaping right before
outputting the preamble.  This simplifies various code.

Fix the escaping in the preamble to only duplicate a leading quote
and leave the rest alone.

Use the shared utility function for quoting macro arguments for
the various internal macros, which eliminates quoting when it's
not necessary.