Force formatting of man refs in L<> in Pod::Man
Man page references in L<> that are detected as such by Pod::Simple are
now always formatted as man page references even if our normal
heuristic would not detect them.  This fixes the formatting of
constructions such as @@RXVT_NAME@@perl(3), which are used by packages
that format a man page with POD and then substitute variables into it
at build time.  Thanks to Marco Sirabella for the analysis and an
initial patch.

Fixes #21