Disable special ` and ' meaning in Pod::Man
Pod::Man now disables the special *roff interpretation of "`" and "'"
characters as paired quotes everywhere, not just in verbatim text, thus
forcing them to be interpreted as the regular ASCII characters.  This
also disables the use of "``" and "''" for paired double-quotes.  The
rationale is similar to that for hyphens: there is no way to tell from
the POD source that the special interpretation as quotes is intended.
To produce paired typographic quotes in the output, use UTF-8 and
Unicode paired quote characters.