Fix wrapping with S<> in Pod::Text subclasses
All Pod::Text subclasses were wrapping on any Unicode whitespace,
which includes non-breaking whitespace that is used if the input
encoding of the POD file is UTF-8. Instead, for now at least, wrap
only on space, tab, and newline. This should be modified to use a
more Unicode-aware algorithm, but that is considerably more work
and I'm not sure how to find that data.

Fixes #24