Update rra-c-util code and minimum Perl version
Update test support libraries and scripts from the forthcoming
rra-c-util 11.0.0 release. This now depends on Perl 5.12 or later,
so update all of podlators to use that same minimum version.
Switch to semantic versioning and the way of declaring package
versions that is supported as of Perl 5.12.

Enabling Perl 5.12 features requires rewriting one of the regexes
in Pod::Man to not strip Unicode spaces such as non-breaking spaces.