Force ragged-right justification with groff 1.23.0
groff 1.23.0 added code to the an macro underlying .IP and .TP that
resets the justification to the AD register, which defaults to b
(full justification). Pod::Man attempts to force ragged-right
justification under nroff with .if n .ad l after .TH, but that change
was reverted in the first =item list, resulting in strange-looking
formatted manual pages in a terminal.

Work around this by setting the AD string to l if the formatter is
nroff in the preamble of each page. This undermines the intended
upstream use of AD as a way for the user to override the formatting,
but Pod::Man-generated man pages already didn't support that and
there currently is no clean way to support the AD string while still
changing the default under nroff to ragged-right justification, as has
been the practice of Pod::Man since its first release in 1999.

Fixes #23