Eliminate 'use autodie' from files touched by Perl core build process
Restore explicit checking for success of system calls as in version 5 of
podlators.

This is needed because podlators is bundled into the Perl core
distribution where it is handled in the build process at a point where
some of autodie's prerequisites are not yet built.

See discussion in https://github.com/rra/podlators/issues/32 and
https://github.com/Perl/perl5/issues/22395.

Co-authored-by: James E Keenan <jkeenan@cpan.org>