install perlpodstyle normally
Move perlpodstyle to lib so it is installed normally. Also remove the
handling from Makefile.PL. It was being installed as a man(1) page, but
section 1 is the wrong location for it. If we don't configure anything
ourselves, ExtUtils::MakeMaker will install the man page in section 3.
This is not perfect, but matches what perl itself does with its
non-library pod files. In the future, perl may understand section 7 and
adjustments can be made for that.