Add none value for Pod::Text encoding option
Pod::Text now supports setting the encoding option to "none" to disable
output encoding. This only makes sense when calling output_string() to
direct the formatted output to a scalar. It is intended for special
cases where the formatted output is held in memory for further
processing and will encoded for output at a later step. Thanks, djerius.

Fixes #38