Work around Freenode::EmptyReturn policy bug
Perl::Critic::Freenode was renamed to Perl::Critic::Community, but
the specific tests were in a weird dual state for a while. In the
`perldocker/perl-tester` images, it still seems to be in that state,
where it reports diagnostics like `Community::EmptyReturn` but
ignores the configuration file exclusion for that rule because
internally it's `Freenode::EmptyReturn`. Excluding the latter in
the configuration file produces a warning that the policy isn't
installed. Excluding it explicitly in `t/style/critic.t` appears to
work around all of this mess.